fix disk/network access issues related with new header.

+++ src/pow.cpp
@@ -81,7 +81,7 @@ bool CheckEquihashSolution(const CBlockHeader *pblock, const CChainParams& param
 
     // I = the block header minus nonce and solution.
     CEquihashInput I{*pblock};
-    int ser_flags = (pblock->nHeight < params.GetConsensus().ContractHeight) ? SERIALIZE_BLOCK_NO_CONTRACT : 0;
+    int ser_flags = (pblock->nHeight < (uint32_t)params.GetConsensus().ContractHeight) ? SERIALIZE_BLOCK_NO_CONTRACT : 0;
     // I||V
     CDataStream ss(SER_NETWORK, PROTOCOL_VERSION | ser_flags);
     ss << I;
